Split forward session modules
