+++ cli/dust-sandbox/src/commands/forward/domain_extract.rs
@@ -1,0 +1,102 @@
+use std::time::Duration;
+
+use tokio::net::TcpStream;
+use tokio::time::{sleep, timeout_at, Instant};
+use tracing::debug;
+
+use super::http_host::parse_http_host;
+use super::sni::parse_client_hello_sni;
+
+const DOMAIN_PEEK_TIMEOUT: Duration = Duration::from_secs(2);
+const DOMAIN_PEEK_RETRY_DELAY: Duration = Duration::from_millis(25);
+const DOMAIN_PEEK_BUFFER_SIZE: usize = 16 * 1024;
+
+#[derive(Debug, PartialEq, Eq)]
+pub(super) enum DomainParseResult {
+    Found(String),
+    NotFound,
+    Incomplete,
+}
+
+#[derive(Debug)]
+pub(super) struct DomainExtraction {
+    pub(super) domain: String,
+    pub(super) failed: bool,
+}
+
+pub(super) fn display_domain(domain: &str) -> &str {
+    if domain.is_empty() {
+        "<unknown>"
+    } else {
+        domain
+    }
+}
+
+pub(super) async fn extract_domain(stream: &TcpStream, original_port: u16) -> DomainExtraction {
+    match original_port {
+        80 => extract_domain_with_parser(stream, parse_http_host).await,
+        443 => extract_domain_with_parser(stream, parse_client_hello_sni).await,
+        _ => DomainExtraction {
+            domain: String::new(),
+            failed: false,
+        },
+    }
+}
+
+async fn extract_domain_with_parser<F>(stream: &TcpStream, parser: F) -> DomainExtraction
+where
+    F: Fn(&[u8]) -> DomainParseResult,
+{
+    let mut buffer = vec![0_u8; DOMAIN_PEEK_BUFFER_SIZE];
+    let deadline = Instant::now() + DOMAIN_PEEK_TIMEOUT;
+
+    loop {
+        let bytes_read = match timeout_at(deadline, stream.peek(&mut buffer)).await {
+            Ok(Ok(bytes_read)) => bytes_read,
+            Ok(Err(error)) => {
+                debug!(error = %error, "failed to peek client bytes for domain extraction");
+                return DomainExtraction {
+                    domain: String::new(),
+                    failed: true,
+                };
+            }
+            Err(_) => {
+                return DomainExtraction {
+                    domain: String::new(),
+                    failed: true,
+                };
+            }
+        };
+
+        if bytes_read == 0 {
+            return DomainExtraction {
+                domain: String::new(),
+                failed: true,
+            };
+        }
+
+        match parser(&buffer[..bytes_read]) {
+            DomainParseResult::Found(domain) => {
+                return DomainExtraction {
+                    domain,
+                    failed: false,
+                };
+            }
+            DomainParseResult::NotFound => {
+                return DomainExtraction {
+                    domain: String::new(),
+                    failed: false,
+                };
+            }
+            DomainParseResult::Incomplete => {
+                if bytes_read == buffer.len() || Instant::now() >= deadline {
+                    return DomainExtraction {
+                        domain: String::new(),
+                        failed: true,
+                    };
+                }
+                sleep(DOMAIN_PEEK_RETRY_DELAY).await;
+            }
+        }
+    }
+}
